+++ tests/fixtures_server.py
@@ -290,7 +290,6 @@ async def MockCurrentUser(app, db):
             default_factory=lambda: ["project"]
         )
         email: Optional[str] = field(default_factory=_random_email)
-        persist: Optional[bool] = True
 
         def _create_user(self):
             defaults = dict(
@@ -312,21 +311,20 @@ async def MockCurrentUser(app, db):
         async def __aenter__(self):
             self._create_user()
 
-            if self.persist:
-                try:
-                    db.add(self.user)
-                    await db.commit()
-                    await db.refresh(self.user)
-                except IntegrityError:
-                    # Safety net, in case of non-unique email addresses
-                    await db.rollback()
-                    self.user.email = _random_email()
-                    db.add(self.user)
-                    await db.commit()
-                    await db.refresh(self.user)
-                # Removing object from test db session, so that we can operate
-                # on user from other sessions
-                db.expunge(self.user)
+            try:
+                db.add(self.user)
+                await db.commit()
+                await db.refresh(self.user)
+            except IntegrityError:
+                # Safety net, in case of non-unique email addresses
+                await db.rollback()
+                self.user.email = _random_email()
+                db.add(self.user)
+                await db.commit()
+                await db.refresh(self.user)
+            # Removing object from test db session, so that we can operate
+            # on user from other sessions
+            db.expunge(self.user)
 
             if self.user.is_active:
                 self.previous_active_user = app.dependency_overrides.get(
@@ -358,8 +356,6 @@ async def MockCurrentUser(app, db):
                 ] = self.current_user_override()
             else:
                 self.previous_active_superuser = None
-
-            return self.user
 
         async def __aexit__(self, *args, **kwargs):